+++ index.html
@@ -230,5 +230,7 @@
     </div>
   </template>
 
+<script src="https://cdn.jsdelivr.net/npm/lodash@4.17.20/lodash.min.js"></script>
+<script src="js/main.js" type="module"></script>
 </body>
 </html>
